+++ src/features/comment/components/LoadingSpinner.tsx
@@ -1,0 +1,14 @@
+"use client";
+
+import React from "react";
+
+export function LoadingSpinner() {
+  return (
+    <div className="min-h-screen bg-gray-50 flex justify-center items-center">
+      <div className="text-center">
+        <div className="animate-spin rounded-full h-8 w-8 border-b-2 border-blue-500 mx-auto mb-4"></div>
+        <p className="text-gray-500">로딩 중...</p>
+      </div>
+    </div>
+  );
+}
